requirements.txt
pytest
selenium
webdriver-manager
pyyaml
